+++ .eslintrc.cjs
@@ -4,8 +4,14 @@ module.exports = {
     es2021: true,
     node: true
   },
-  plugins: ['@typescript-eslint', 'jsdoc', 'eslint-plugin-tsdoc', 'prettier'],
-  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended', 'plugin:prettier/recommended'],
+  plugins: ['@typescript-eslint', 'jsdoc', 'eslint-plugin-tsdoc', 'prettier', 'import'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
+    'plugin:import/recommended',
+    'plugin:import/typescript'
+  ],
   parser: '@typescript-eslint/parser',
   overrides: [],
   parserOptions: {
@@ -90,6 +96,18 @@ module.exports = {
     'jsdoc/valid-types': 1, // Recommended
     'tsdoc/syntax': 'warn',
     // Prettier plugin rules
-    'prettier/prettier': 'error'
+    'prettier/prettier': 'error',
+    // Import order plugin rules
+    'import/order': [
+      'error',
+      {
+        'newlines-between': 'always',
+        alphabetize: {
+          order: 'asc' /* sort in ascending order. Options: ['ignore', 'asc', 'desc'] */,
+          caseInsensitive: true /* ignore case. Options: [true, false] */
+        },
+        groups: ['builtin', 'external', 'parent', 'sibling', 'index']
+      }
+    ]
   }
 }