+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Changelog for the `minimal-ts` package.
 
+## 1.4.0 TBD
+
+- ci: github workflow action pre-commit, use lint-staged
+- chore: use prettier eslint plugin
+- chore: use import order eslint plugin
+- chore: upgraded dependencies
+
 ## 1.3.5
 
 - added github actions CI/CD